Removed PHP support altogether
I seems that support for PHP is something that is not possible in a way
we at pilight feel comfortable with. There is simply no easy way to
interface with PHP through C library calls so unstable php-cgi
system calls where required.

Until PHP allows is to be embedded easily we will refrain support.