Fix buffer overflow in setenv() and unsetenv() functions.
A buffer allocated by setenv() and unsetenv() was one byte too short.
If setenv() was called with option to not overwrite an existing variable
and that variable does not exist, then a NULL pointer was used and a
segfault happens (if the variable exists, it was unnecessary set again
to its current value). All this has been fixed.

Further the parameters passed to setenv() and unsetenv() are now checked
against NULL pointers and the functions do the best they can do without
crashing (return with failure indicator and errno set). This introduced
a feature for setenv() that, if called with a variable name but a value
of NULL, it does an unsetenv().