Webgui: Fix dimmer display bugs
This would cause the text field next to the slider to be
initialized to 0 at page load, even though the slider was showing the
correct dimlevel.

This commit fixes this by also assigning the dimlevel to the text field on
page load. Patch by @Niek.

Further, the slider-input field was editable, but without use (editing the
value has no effect). Therefore, display the field as disabled.