Make json type defines useable as bitmask
The former json types could not easily be used as bitmask, because the
values had some bits in common. For example JSON_STRING (0x04) and
JSON_ARRAY (0x16) had both the third bit set to true. So a calue
defined as JSON_ARRAY had internally the same value as
(JSON_STRING | JSON_ARRAY).

This is a problem, since for example teh generic_label protocol uses
bitwise or to declare a parameter as "multiple types" (in this case its
JSON_STRING | JSON_NUMBER.

Tgis uses bitshifts to define differentiable values.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>